menu-frontend/backend/requirements.txt
flask
flask_cors
pandas
python-dotenv
openai==1.55.3
httpx==0.27.2
gunicorn
